categories and product use case refactored
