+++ app/src/main/java/com/fakestore/domain/CategoriesUseCase.kt
@@ -1,18 +1,8 @@
 package com.fakestore.domain
 
-import com.fakestore.data.remote.FakestoreRepository
+
 import com.fakestore.domain.model.Categories
-import retrofit2.Response
 
-object CategoriesUseCase {
-    suspend fun getCategories(): Categories {
-        val repo = FakestoreRepository
-        val categoriesFromRepo = repo.fakestoreApi.getCategories()
-        return convertCategories(categoriesFromRepo)
-
-    }
-
-    private fun convertCategories(categories: Response<List<String>>): Categories {
-        return Categories(categories.body() ?: listOf())
-    }
+interface CategoriesUseCase {
+    suspend fun getCategories(): Categories
 }
