+++ app/src/main/java/com/fakestore/ui/home/HomeFragment.kt
@@ -10,6 +10,8 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
+import com.fakestore.data.remote.CategoriesUseCaseImpl
+import com.fakestore.data.remote.ProductUseCaseImpl
 import com.fakestore.databinding.FragmentHomeBinding
 import com.fakestore.domain.CategoriesUseCase
 import com.fakestore.domain.ProductUseCase
@@ -21,10 +23,22 @@ class HomeFragment : Fragment() {
 
     init {
         lifecycleScope.launch {
-            val response = ProductUseCase.getProduct()
+            // TODO val productUseCase: ProductUseCase = object ProductUseCaseImpl
+            // не сработало
+
+            //  val response = ProductUseCase.getProduct()
+
+            val productUseCase: ProductUseCase = object : ProductUseCaseImpl() {}
+
             Log.d("response", Calendar.getInstance().time.toString())
-            Log.d("response", response.toString())
-            val categories = CategoriesUseCase.getCategories()
+            Log.d("response", productUseCase.toString())
+
+
+            // val categories = CategoriesUseCase.getCategories()
+            val categoriesUseCase: CategoriesUseCase = object : CategoriesUseCaseImpl() {}
+
+            Log.d("response", Calendar.getInstance().time.toString())
+            Log.d("response", categoriesUseCase.getCategories().categories.toString())
         }
     }
 
@@ -50,9 +64,4 @@ class HomeFragment : Fragment() {
             textView.text = it
         }
     }
-    //TODO спросить про  binding = null и View?
-//    override fun onDestroyView() {
-//        super.onDestroyView()
-//        binding = null
-//    }
 }
