+++ app/src/main/java/com/fakestore/ui/home/HomeViewModel.kt
@@ -15,6 +15,7 @@ class HomeViewModel : ViewModel() {
             homeString.postValue(getNotifications())
 
             // дернуть юзкейс
+            //
         }
     }
