+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -18,5 +18,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+
+        android:background="@android:color/holo_blue_bright" />
 </androidx.constraintlayout.widget.ConstraintLayout>
