+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -9,7 +9,10 @@
         android:id="@+id/navigation_home"
         android:name="com.fakestore.ui.home.HomeFragment"
         android:label="@string/title_home"
-        tools:layout="@layout/fragment_home" />
+        tools:layout="@layout/fragment_home">
+
+
+    </fragment>
 
     <fragment
         android:id="@+id/navigation_search"
